Implement connection forwarding and timeouts
For now, we just check if the server is reachable and if so just forward
all data between client and server.
This makes Portal functional as a simple proxy.

Currently, we give each paket up to 5s to arrive, which should be a
reasonable default for now.
Note that this also introduces another error type for timeouts.
Hence we add a custom error type to Portal, which will eventually contain
all kinds of error we encounter.